refactor: Restructure project with backend package and environment-specific configs
This commit introduces a significant refactoring of the project structure:
- All core application classes (controllers, services, entities, enums, repositories) have been moved into a new  package within .
- Environment-specific application property files (, , ) have been added to support different configurations for various deployment stages.
- The  has been updated to include the H2 database dependency, primarily for use in the test environment.
- The main  now uses environment variables for database configuration and activates profiles dynamically.

These changes aim to improve project organization, modularity, and facilitate easier configuration management across different environments.
