+++ src/main/java/com/wizeflow/crm/backend/WizeFlowCrmApplication.java
@@ -1,4 +1,4 @@
-package com.wizeflow.crm;
+package com.wizeflow.crm.backend;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
