+++ src/main/java/com/wizeflow/crm/backend/enums/ChatSessionChannel.java
@@ -1,4 +1,4 @@
-package com.wizeflow.crm_backend.enums;
+package com.wizeflow.crm.backend.enums;
 
 public enum ChatSessionChannel {
     WHATSAPP,
