+++ src/main/java/com/wizeflow/crm/backend/infrastructure/entity/ChatSession.java
@@ -1,6 +1,6 @@
-package com.wizeflow.crm_backend.infrastructure.entity;
+package com.wizeflow.crm.backend.infrastructure.entity;
 
-import com.wizeflow.crm_backend.enums.ChatSessionChannel;
+import com.wizeflow.crm.backend.enums.ChatSessionChannel;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
