+++ src/main/java/com/wizeflow/crm/backend/infrastructure/entity/Lead.java
@@ -1,6 +1,6 @@
-package com.wizeflow.crm_backend.infrastructure.entity;
+package com.wizeflow.crm.backend.infrastructure.entity;
 
-import com.wizeflow.crm_backend.enums.Source;
+import com.wizeflow.crm.backend.enums.Source;
 import jakarta.persistence.*;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
