+++ src/main/resources/application-dev.properties
@@ -1,0 +1,1 @@
+spring.application.name=WizeFlow CRM
