+++ src/main/resources/application.properties
@@ -1,1 +1,10 @@
-spring.application.name=WizeFlow CRM
+spring.profiles.active=${SPRING_PROFILES_ACTIVE:dev}
+
+# -- Datasource (PostgreSQL) --
+spring.datasource.url=${DB_URL}
+spring.datasource.username=${DB_USERNAME}
+spring.datasource.password=${DB_PASSWORD}
+spring.datasource.driver-class-name=org.postgresql.Driver
+
+# -- JPA / Hibernate --
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
