+++ src/test/java/com/wizeflow/crm/backend/WizeFlowCrmApplicationTests.java
@@ -1,9 +1,11 @@
-package com.wizeflow.crm;
+package com.wizeflow.crm.backend;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 
 @SpringBootTest
+@ActiveProfiles("test")
 class WizeFlowCrmApplicationTests {
 
 	@Test
